remove and update validation

+++ application/src/domain/services/TreasuryCashDomainService.js
@@ -147,30 +147,42 @@ export class TreasuryCashDomainService {
                 ? this._paymentValidate(entity) : this._receiveValidate(entity),
             journalId = entity.journalId;
 
+        if (persistedTreasury.journalId)
+            errors.push('برای {0} نقدی سند صادر شده است و امکان ویرایش وجود ندارد!'
+                .format(Enums.TreasuryType().getDisplay(persistedTreasury.treasuryType)));
+
         if (errors.length > 0)
             throw new ValidationException(errors);
 
 
-        entity.journalId = null;
+        //entity.journalId = null;
         this.treasuryRepository.update(id, entity);
 
-        if (journalId) {
+/*        if (journalId) {
             this.journalDomainService.remove(journalId);
 
             entity.treasuryType === 'receive' ?
                 this.treasuryJournalGenerationDomainService.generateForReceiveCash(id) :
                 this.treasuryJournalGenerationDomainService.generateForPaymentCash(id);
-        }
+        }*/
 
     }
 
     remove(id) {
-        let persistedTreasury = this.treasuryRepository.findById(id);
+        let persistedTreasury = this.treasuryRepository.findById(id),
+            errors = [];
+
+        if (persistedTreasury.journalId)
+            errors.push('برای {0} نقدی سند صادر شده است و امکان حذف وجود ندارد!'
+                .format(Enums.TreasuryType().getDisplay(persistedTreasury.treasuryType)));
+
+        if (errors.length > 0)
+            throw new ValidationException(errors);
 
         this.treasuryRepository.remove(id);
 
-        if (persistedTreasury.journalId)
-            this.journalDomainService.remove(persistedTreasury.journalId);
+        /*if (persistedTreasury.journalId)
+            this.journalDomainService.remove(persistedTreasury.journalId);*/
     }
